The Windows functions intpath2winpath and win_utf8ncasecmp now verifies that the input is valid UTF-8. Without this, MultiByteToWideChar will silently skip invalid sequences, which might result in a different file name. intpath2winpath returns NULL if input is not valid UTF-8.
All callers of intpath2winpath now checks if the result is NULL, and
returns EINVAL in this case. This is necessary in other cases as well,
for example, when out of memory, or if MultiByteToWideChar fails for
some reason.

error.c has been adapted to translate EINVAL to NFS3ERR_INVAL (which
is the best NFSv3 error code we have to indicate an invalid file name)
in a few places.


git-svn-id: svn://svn.code.sf.net/p/unfs3/code/unfs3.svn/trunk@387 fb7659fb-8760-4330-86f7-fdbeb49e81a3