daemon.c: Fix race window for writing of the pid file
The parent process should write the pid file such that the pid file
will can be checked immediately following exit of the fork from the
parent.

This allows external monitoring applications to watch the daemon
without having to add sleep calls to wait for the pid file be written
on a busy system.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>