nfs.c: Allow max sa.sun_path for a localdomain socket with the user nfs-server
There is a hard limit for the kernel of 108 characters for a
localdomain socket name.  To avoid problems with the user nfs
server it should maximize the number of characters by using
a relative path on the server side.

Previously the nfs-server used the absolute path name passed to
the sa.sunpath arg for binding the socket and this has caused
problems for both the X server and UST binaries which make
heavy use of named sockets with long names.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>