Handle directories without mtime/ctime
Some directories might not store mtime/ctime, like the root directory
of FAT. However it is critical for NFS that these fields are updated
when files are added or removed.

Work around this by tracking the contents of the directories and using
the current time as the mtime/ctime whenever we notice a change.