attr: fix utime for symlink
unfs3 has an old defect that it can not change the timestamps of a
symlink file because it only uses utime(), which will follow the
symlink. This will not cause an error if the symlink points to an
existent file. But under some special situation, such as installing
a rpm package, rpm tool will create the symlink first and try to
modify the timestamps of it, when the target file is non-existent.
This will cause an ESTALE error. Making rpm tool ignore this error
is a solution, but not the best one. An acceptable approach is
Making unfs3 support lutimes(), which can modify the symlink file
itself. Considering not every system support this function, so a
function checking is necessary.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>