Fix up various style issues
Our indent-all script has not been used in ages, so the style has been
allowed to drift. Running it again results in massive changes, partly
because the behaviour of "indent" has changed over the years.

The "indent" command seems to be a bit too opinionated for our usage as
it wants a say in everything. Let's switch to astyle instead, which only
adjusts certain things, allowing more variation in areas we don't want
to enforce.

We've excluded xdr.c since it is automatically generated.